+++ backend/core/src/seed.ts
@@ -28,6 +28,8 @@ import { Listing10145Seed } from "./seeds/listings/listing-detroit-10145"
 import { CountyCode } from "./shared/types/county-code"
 import { ListingTreymoreSeed } from "./seeds/listings/listing-detroit-treymore"
 import { UserRoles } from "./auth/entities/user-roles.entity"
+import { AmiChart } from "./ami-charts/entities/ami-chart.entity"
+import { WayneCountyMSHDA2021 } from "./seeds/ami-charts"
 import { ListingDefaultMultipleAMI } from "./seeds/listings/listing-default-multiple-ami"
 import { ListingDefaultMultipleAMIAndPercentages } from "./seeds/listings/listing-default-multiple-ami-and-percentages"
 
@@ -177,6 +179,10 @@ async function seed() {
       }
     }
   }
+
+  // Seed the Detroit AMI data, since it's not linked to any units.
+  const amiChartRepo = app.get<Repository<AmiChart>>(getRepositoryToken(AmiChart))
+  await amiChartRepo.save(JSON.parse(JSON.stringify(WayneCountyMSHDA2021)))
 
   await userRepo.save(admin)
   const roles: UserRoles = { user: admin, isPartner: true, isAdmin: true }